Updated Purchase, PurchaseSQLDAO, Purchase was returning a double instead of an int, PurchaseSQLDAO is now able to substractUnits from an instance

+++ src/Business/Entities/Purchase.java
@@ -12,6 +12,6 @@ public class Purchase {
     }
 
     public String getCrypto() { return crypto; }
-    public double getUnits() { return units; }
+    public int getUnits() { return units; }
     public double getPriceUnit() { return priceUnit; }
 }
